Sending caps key state to puppet keyboard
We want the lighting effect to change when caps key is activated.
The caps key state is not usually sent to the puppet keyboard. (Probably
because it does not need to, to make the feature work.)

This commit adds a custom rpc message to send one byte to the puppet,
containing the current state of the caps key feature.